refactor: mission detail 파일 옮기기
