--- src/app/mission/[id]/detail/layout.tsx
@@ -1,15 +1,0 @@
-import { type PropsWithChildren } from 'react';
-import { css } from '@styled-system/css';
-
-export default function Layout({ children }: PropsWithChildren) {
-  return <div className={containerCss}>{children}</div>;
-}
-
-const containerCss = css({
-  maxWidth: 'maxWidth',
-  margin: '0 auto',
-  minHeight: '100vh',
-
-  display: 'flex',
-  flexDirection: 'column',
-});
