+++ src/pages/mission/[id]/detail/MissionHistoryBanner/MissionHistoryBanner.tsx
