+++ src/pages/mission/[id]/detail/MissionHistoryBanner/MissionHistoryBannerApi.tsx
@@ -1,7 +1,7 @@
 import { useGetMissionDetailNoSuspense } from '@/apis/mission';
-import MissionHistorySkeleton from '@/app/mission/[id]/detail/MissionHistoryBanner/MissionHistorySkeleton';
 import Banner from '@/components/Banner/Banner';
 import { MISSION_CATEGORY_LABEL } from '@/constants/mission';
+import MissionHistorySkeleton from '@/pages/mission/[id]/detail/MissionHistoryBanner/MissionHistorySkeleton';
 
 function MissionHistoryBannerApi({ missionId }: { missionId: string }) {
   const { data } = useGetMissionDetailNoSuspense(missionId);
